Add initial support for mender external env
Add initial support for a mender environment that is stored on
an ext4 filesystem rather than with the U-Boot environment. The
difference between this and storing the entire U-Boot environment
on an ext4 filesystem is that U-Boot can use an embedded environment
(i.e. CONFIG_ENV_IS_NOWHERE=y) for most of its environment and only
pull the mender-relevant values from the external environment.

Signed-off-by: Thomas Benson <tbensongit@gmail.com>