Do not force U-Boot default env for USB boots
When booting via USB, the default environment is applied after the
mender external environment overlay. Since in our case, the main U-Boot
environment is not stored externally, we already loaded the default
environment. The only allowed changed would be if we successfully
loaded the mender external environment, which we do not want to
override.

Signed-off-by: Thomas Benson <tbensongit@gmail.com>